chore(DnD): DRY out common utility fn
