+++ src/addons/dragAndDrop/common.js
@@ -1,5 +1,6 @@
 import { wrapAccessor } from '../../utils/accessors'
 import { createFactory } from 'react'
+import * as dates from '../../utils/dates'
 
 export const dragAccessors = {
   start: wrapAccessor(e => e.start),
@@ -30,4 +31,16 @@ export function pointInColumn(bounds, point) {
   const { left, right, top } = bounds
   const { x, y } = point
   return x < right + 10 && x > left && y > top
+}
+
+export function eventTimes(event, accessors) {
+  let start = accessors.start(event)
+  let end = accessors.end(event)
+
+  const isZeroDuration =
+    dates.eq(start, end, 'minutes') && start.getMinutes() === 0
+  // make zero duration midnight events at least one day long
+  if (isZeroDuration) end = dates.add(end, 1, 'day')
+  const duration = dates.diff(end, start, 'milliseconds')
+  return { start, end, duration }
 }
